feat: move to open api spec
